+++ src/main/java/baseball/Application.java
@@ -1,6 +1,6 @@
 package baseball;
 
-import baseball.core.BaseballGame;
+import baseball.game.BaseballGame;
 
 public class Application {
     public static void main(String[] args) {
